Move common methods into Collections::Base
There were quite a few methods that were duplicated across the
collection classes for selects, radio buttons and checkbox groups, so
I've introduced a new `Collections::Base` class to hold those.
