+++ lib/citizens_advice_form_builder/elements.rb
@@ -7,6 +7,7 @@ require_relative "elements/text_area"
 require_relative "elements/date_input"
 require_relative "elements/error_summary"
 
+require_relative "elements/collections/base"
 require_relative "elements/collections/radio_buttons"
 require_relative "elements/collections/check_boxes"
 require_relative "elements/collections/select"
